Replace mambajs solver on js-rattler-solver
