--- src/conda-packages-solver/core.wasm.d.ts
@@ -1,2 +1,0 @@
-declare const coreWasm: any;
-export default coreWasm;
